extracted input handler
