+++ src/main/java/Duke.java
@@ -4,7 +4,6 @@ import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.Scanner;
 
 public class Duke {
@@ -36,56 +35,6 @@ public class Duke {
             System.out.println("    ------------------------------------------");
             System.out.println("    " + input);
             System.out.println("    ------------------------------------------");
-        }
-    }
-
-    static class InputHandler {
-        private final Scanner scanner = new Scanner(System.in);
-        private final Model model;
-        private final TaskStorage taskStorage = new TaskStorage();
-
-        private boolean isExit = false;
-
-        public InputHandler(Model model) {
-            this.model = model;
-        }
-
-        // TODO: Implement a parser instead of hard coding if-elses here
-        public void handleInput() {
-            String input = scanner.nextLine();
-            if (input.equals("bye")) {
-                isExit = true;
-                return;
-            }
-
-            if (input.equals("list")) {
-                System.out.println(taskStorage);
-                return;
-            }
-
-            if (input.startsWith("mark")) {
-                int index = Integer.parseInt(input.split(" ")[1]) - 1;
-                model.echo(taskStorage.markAsDone(index));
-                return;
-            }
-
-            if (input.startsWith("unmark")) {
-                int index = Integer.parseInt(input.split(" ")[1]) - 1;
-                model.echo(taskStorage.unmarkAsDone(index));
-                return;
-            }
-
-            if (input.startsWith("delete")) {
-                int index = Integer.parseInt(input.split(" ")[1]) - 1;
-                model.echo(taskStorage.delete(index));
-                return;
-            }
-
-            model.echo(taskStorage.save(input));
-        }
-
-        public boolean isExit() {
-            return isExit;
         }
     }
